@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { SparklineComponent, Inject, SparklineTooltip } from '@syncfusion/ej2-react-charts';
 
 class SparkLine extends React.PureComponent {
@@ -16,7 +17,6 @@ class SparkLine extends React.PureComponent {
         border={{ color: currentColor, width: 2 }}
         tooltipSettings={{
           visible: true,
-          // eslint-disable-next-line no-template-curly-in-string
           format: '${x} : data ${yval}',
           trackLineSettings: {
             visible: true,
@@ -34,4 +34,26 @@ class SparkLine extends React.PureComponent {
   }
 }
 
+SparkLine.propTypes = {
+  id: PropTypes.string.isRequired,
+  height: PropTypes.string.isRequired,
+  width: PropTypes.string.isRequired,
+  color: PropTypes.string.isRequired,
+  data: PropTypes.arrayOf(
+    PropTypes.shape({
+      x: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+      yval: PropTypes.number.isRequired,
+    })
+  ).isRequired,
+  type: PropTypes.string.isRequired,
+  currentColor: PropTypes.string.isRequired,
+};
+
+SparkLine.defaultProps = {
+  color: 'blue',
+  type: 'Line',
+  currentColor: 'black',
+};
+
 export default SparkLine;
+
