@@ -1,7 +1,8 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 const ChartsHeader = ({ category, title }) => (
-  <div className=" mb-10">
+  <div className="mb-10">
     <div>
       <p className="text-lg text-gray-400">Chart</p>
       <p className="text-3xl font-extrabold tracking-tight dark:text-gray-200 text-slate-900">{category}</p>
@@ -10,4 +11,10 @@ const ChartsHeader = ({ category, title }) => (
   </div>
 );
 
+ChartsHeader.propTypes = {
+  category: PropTypes.string.isRequired,
+  title: PropTypes.string.isRequired,
+};
+
 export default ChartsHeader;
+
