@@ -1,17 +1,25 @@
-export { default as Button } from './Button';
-export { default as ThemeSettings } from './ThemeSettings';
-export { default as Sidebar } from './Sidebar';
-// eslint-disable-next-line import/no-cycle
-export { default as Navbar } from './Navbar';
-export { default as Footer } from './Footer';
-export { default as Cart } from './Cart';
-export { default as Chat } from './Chat';
-export { default as Notification } from './Notification';
-export { default as UserProfile } from './UserProfile';
-export { default as SparkLine } from './Charts/SparkLine';
-export { default as LineChart } from './Charts/LineChart';
-export { default as Stacked } from './Charts/Stacked';
-export { default as Pie } from './Charts/Pie';
-export { default as ChartsHeader } from './ChartsHeader';
-export { default as Header } from './Header';
+// index.js
+
+// Exporting components directly to avoid potential cyclic dependencies
+export { default as Button } from './Button/Button';
+export { default as ThemeSettings } from './ThemeSettings/ThemeSettings';
+export { default as Sidebar } from './Sidebar/Sidebar';
+export { default as Navbar } from './Navbar/Navbar';
+export { default as Footer } from './Footer/Footer';
+export { default as Cart } from './Cart/Cart';
+export { default as Chat } from './Chat/Chat';
+export { default as Notification } from './Notification/Notification';
+export { default as UserProfile } from './UserProfile/UserProfile';
+
+// Exporting chart components
+export { default as SparkLine } from './Charts/SparkLine/SparkLine';
+export { default as LineChart } from './Charts/LineChart/LineChart';
+export { default as Stacked } from './Charts/Stacked/Stacked';
+export { default as Pie } from './Charts/Pie/Pie';
+
+// Exporting additional components
+export { default as ChartsHeader } from './ChartsHeader/ChartsHeader';
+export { default as Header } from './Header/Header';
+
+
 
